added mysql file and renamed db to db_mongo.ts
