+++ package.json
@@ -48,6 +48,7 @@
     "lodash": "^4.17.21",
     "mockdate": "^3.0.5",
     "mongodb": "^4.6.0",
+    "mysql2": "^2.3.3",
     "pino": "^7.11.0",
     "prettier": "^2.6.2",
     "superagent": "^6.1.0",
